+++ src/index.ts
@@ -49,3 +49,4 @@ export {
 export * from './components/Text';
 export { default as UserManagement, type UserManagementProps } from './components/UserManagement/UserManagement';
 export * from './types/commonTypes';
+export { Breadcrumb, type BreadcrumbProps } from './components/Breadcrumb';